[PR] 장소 검색한 아이템 클릭 시 viewModel 저장해서 MapFragment에서 관측 후 마커 및 패널 이동 구현

+++ data/src/main/java/com/stop/data/remote/network/NearPlaceApiService.kt
@@ -11,8 +11,8 @@ interface NearPlaceApiService {
     suspend fun getNearPlaces(
         @Query("version") version: Int,
         @Query("searchKeyword") searchKeyword: String,
-        @Query("centerLon") centerLon: Float,
-        @Query("centerLat") centerLat: Float,
+        @Query("centerLon") centerLon: Double,
+        @Query("centerLat") centerLat: Double,
         @Query("appKey") appKey: String,
     ): NetworkResult<NearPlcaeResponse>
