+++ data/src/main/java/com/stop/data/remote/source/nearplace/NearPlaceRemoteDataSource.kt
@@ -7,8 +7,8 @@ interface NearPlaceRemoteDataSource {
     suspend fun getNearPlaces(
         version : Int,
         searchKeyword: String,
-        centerLon: Float,
-        centerLat: Float,
+        centerLon: Double,
+        centerLat: Double,
         appKey: String
     ): List<Place>
